extract countdown out of dashboard
